% Generated by roxygen2 (4.0.2): do not edit by hand
\name{MOA_classification_bayes}
\alias{MOA_classification_bayes}
\alias{NaiveBayes}
\alias{NaiveBayesMultinomial}
\title{MOA bayesian classification}
\usage{
NaiveBayes(control = NULL, ...)

NaiveBayesMultinomial(control = NULL, ...)
}
\arguments{
\item{control}{an object of class \code{MOAmodelOptions} as obtained by calling \code{\link{MOAoptions}}}

\item{...}{options of parameters passed on to \code{\link{MOAoptions}}, in case \code{control} is left to NULL.
Ignored if \code{control} is supplied}
}
\value{
An object of class \code{MOA_classifier} which sets up an untrained MOA model,
which can be trained using \code{\link{trainMOA}}
}
\description{
MOA bayesian classification
}
\examples{
ctrl <- MOAoptions(model = "NaiveBayes")
mymodel <- NaiveBayes(control=ctrl)
mymodel
}
\seealso{
\code{\link{MOAoptions}}, \code{\link{trainMOA}}
}

