% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/Recommenders.R
\name{MOA_recommendation_engines}
\alias{BRISMFPredictor}
\alias{BaselinePredictor}
\alias{MOA_recommendation_engines}
\title{MOA recommendation engines}
\usage{
BRISMFPredictor(control = NULL, ...)

BaselinePredictor(control = NULL, ...)
}
\arguments{
\item{control}{an object of class \code{MOAmodelOptions} as obtained by calling \code{\link{MOAoptions}}}

\item{...}{options of parameters passed on to \code{\link{MOAoptions}}, in case \code{control} is left to NULL.
Ignored if \code{control} is supplied}
}
\value{
An object of class \code{MOA_recommender} which sets up an untrained MOA model,
which can be trained using \code{\link{trainMOA}}
}
\description{
MOA recommendation engines
}
\examples{
ctrl <- MOAoptions(model = "BRISMFPredictor", features = 10)
brism <- BRISMFPredictor(control=ctrl)
brism
baseline <- BaselinePredictor()
baseline
}
\seealso{
\code{\link{MOAoptions}}, \code{\link{trainMOA}}
}

