% Generated by roxygen2 (4.0.2): do not edit by hand
\name{MOA_classification_trees}
\alias{ASHoeffdingTree}
\alias{AdaHoeffdingOptionTree}
\alias{DecisionStump}
\alias{HoeffdingAdaptiveTree}
\alias{HoeffdingOptionTree}
\alias{HoeffdingTree}
\alias{LimAttHoeffdingTree}
\alias{MOA_classification_trees}
\alias{RandomHoeffdingTree}
\title{MOA classification trees}
\usage{
AdaHoeffdingOptionTree(control = NULL, ...)

ASHoeffdingTree(control = NULL, ...)

DecisionStump(control = NULL, ...)

HoeffdingAdaptiveTree(control = NULL, ...)

HoeffdingOptionTree(control = NULL, ...)

HoeffdingTree(control = NULL, ...)

LimAttHoeffdingTree(control = NULL, ...)

RandomHoeffdingTree(control = NULL, ...)
}
\arguments{
\item{control}{an object of class \code{MOAmodelOptions} as obtained by calling \code{\link{MOAoptions}}}

\item{...}{options of parameters passed on to \code{\link{MOAoptions}}, in case \code{control} is left to NULL.
Ignored if \code{control} is supplied}
}
\value{
An object of class \code{MOA_classifier} which sets up an untrained MOA model,
which can be trained using \code{\link{trainMOA}}
}
\description{
MOA classification trees
}
\examples{
ctrl <- MOAoptions(model = "HoeffdingTree", leafprediction = "MC",
   removePoorAtts = TRUE, binarySplits = TRUE, tieThreshold = 0.20)
hdt <- HoeffdingTree(control=ctrl)
hdt
hdt <- HoeffdingTree(numericEstimator = "GaussianNumericAttributeClassObserver")
hdt
}
\seealso{
\code{\link{MOAoptions}}, \code{\link{trainMOA}}
}

