% Generated by roxygen2 (4.0.1): do not edit by hand
\name{MOA_recommender}
\alias{MOA_recommender}
\title{Create a MOA classifier}
\usage{
MOA_recommender(model, control = NULL, ...)
}
\arguments{
\item{model}{character string with a model.
E.g. HoeffdingTree, DecisionStump, NaiveBayes, HoeffdingOptionTree, ...
The list of known models can be obtained by typing RMOA:::.moaknownmodels.
See the examples and \code{\link{MOAoptions}}.}

\item{control}{an object of class \code{MOAmodelOptions} as obtained by calling \code{\link{MOAoptions}}}

\item{...}{options of parameters passed on to \code{\link{MOAoptions}}, in case \code{control} is left to NULL.
Ignored if \code{control} is supplied}
}
\value{
An object of class \code{MOA_classifier}
}
\description{
Create a MOA classifier
}
\examples{
RMOA:::.moaknownmodels
ctrl <- MOAoptions(model = "BRISMFPredictor", features = 10, lRate=0.002)
brism <- MOA_recommender(model = "BRISMFPredictor", control=ctrl)
brism
MOAoptions(model = "BaselinePredictor")
baseline <- MOA_recommender(model = "BaselinePredictor")
baseline
}
\seealso{
\code{\link{MOAoptions}}
}

