% Generated by roxygen2 (4.0.2): do not edit by hand
\name{summary.MOA_classifier}
\alias{summary.MOA_classifier}
\title{Summary statistics of a MOA classifier}
\usage{
\method{summary}{MOA_classifier}(object, ...)
}
\arguments{
\item{object}{an object of class  \code{MOA_classifier}}

\item{...}{other arguments, currently not used yet}
}
\value{
the form of the return value depends on the type of MOA model
}
\description{
Summary statistics of a MOA classifier
}
\examples{
hdt <- HoeffdingTree(numericEstimator = "GaussianNumericAttributeClassObserver")
hdt
data(iris)
iris <- factorise(iris)
irisdatastream <- datastream_dataframe(data=iris)
## Train the model
hdtreetrained <- trainMOA(model = hdt,
 Species ~ Sepal.Length + Sepal.Width + Petal.Length + Petal.Width,
 data = irisdatastream)
summary(hdtreetrained$model)
}

