% Generated by roxygen2 (4.0.2): do not edit by hand
\name{summary.MOA_regressor}
\alias{summary.MOA_regressor}
\title{Summary statistics of a MOA regressor}
\usage{
\method{summary}{MOA_regressor}(object, ...)
}
\arguments{
\item{object}{an object of class  \code{MOA_regressor}}

\item{...}{other arguments, currently not used yet}
}
\value{
the form of the return value depends on the type of MOA model
}
\description{
Summary statistics of a MOA regressor
}
\examples{
## TODO
}

