% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/train.R
\name{trainMOA}
\alias{trainMOA}
\title{Train a MOA classifier/regressor/recommendation engine on a datastream}
\usage{
trainMOA(model, ...)
}
\arguments{
\item{model}{an object of class \code{MOA_model}, as returned by \code{\link{MOA_classifier}},
\code{\link{MOA_regressor}}, \code{\link{MOA_recommender}}}

\item{...}{other parameters passed on to the methods}
}
\value{
An object of class MOA_trainedmodel which is returned by the methods for the specific model.
See \code{\link{trainMOA.MOA_classifier}}, \code{\link{trainMOA.MOA_regressor}}, \code{\link{trainMOA.MOA_recommender}}
}
\description{
Train a MOA classifier/regressor/recommendation engine on a datastream
}
\seealso{
\code{\link{trainMOA.MOA_classifier}}, \code{\link{trainMOA.MOA_regressor}}, \code{\link{trainMOA.MOA_recommender}}
}

